back-end/requirements.txt
flashtorch==0.1.3
Flask==2.0.1
importlib_resources==5.2.2
lib==3.0.0
matplotlib==3.4.2
numpy==1.21.2
Pillow==8.4.0
pytorch_influence_functions==0.1.1
torch==1.9.1
torchattacks==3.1.0
torchvision==0.10.1
webunit==1.3.10
Werkzeug==2.0.1
